[style] modify function to util.js
